+++ packages/shared/sdk-client/src/configuration/Configuration.ts
@@ -12,7 +12,7 @@ import {
 import { Hook, type LDOptions } from '../api';
 import { LDContext } from '../api/LDContext';
 import { LDInspection } from '../api/LDInspection';
-import type { LDDebugOverride } from '../flag-manager/FlagManager';
+import type { LDDebugOverride } from '../api/LDDebugOverride';
 import type {
   InternalDataSystemOptions,
   PlatformDataSystemDefaults,
